[add] Errorの型を追加
