+++ core/domain/src/main/kotlin/com/wsr/exceptions/GetAllException.kt
@@ -1,0 +1,5 @@
+package com.wsr.exceptions
+
+sealed class GetAllException : Throwable() {
+    class DatabaseException : GetAllException()
+}
