+++ core/usecase/src/main/kotlin/com/wsr/passwordgroup/GetAllPasswordGroupUseCase.kt
@@ -3,7 +3,7 @@ package com.wsr.passwordgroup
 import com.wsr.state.State
 import kotlinx.coroutines.flow.StateFlow
 
-interface GetPasswordGroupUseCase {
+interface GetAllPasswordGroupUseCase {
     val data: StateFlow<State<List<PasswordGroupUseCaseModel>, Throwable>>
 
     suspend fun getAllByEmail(email: String)
